fix adding to map order
was adding the current item to the internal cache before the actual hash
collision checking
